+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Supported customized request headers for AI services. (#301)
+
 ## [1.3.8] - 2025-12-24
 
 ### Added